Added duecredit for remaining `fatiando.gravmag`  modules
Note that some are currently broken due to a bug in citeproc-py, in
particular where page numbers are non-integers.

This is now complete for the `gravmag` module, although I noticed some
modules do not have references in the docstrings of individual methods.
If desired, they will be fixed as part of this PR.